sort of cmd files
