+++ cmd/kubeless/autoscale/autoscaleList.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package autoscale
 
 import (
 	"encoding/json"
