+++ cmd/kubeless/autoscale/autoscaleList_test.go
@@ -1,4 +1,4 @@
-package main
+package autoscale
 
 import (
 	"bytes"
