+++ cmd/kubeless/function/delete.go
@@ -1,9 +1,12 @@
 /*
 Copyright (c) 2016-2017 Bitnami
+
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
+
     http://www.apache.org/licenses/LICENSE-2.0
+
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -11,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package function
 
 import (
 	"github.com/kubeless/kubeless/pkg/utils"
@@ -19,15 +22,15 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var routeDeleteCmd = &cobra.Command{
-	Use:   "delete <name>",
-	Short: "delete a route from Kubeless",
-	Long:  `delete a route from Kubeless`,
+var deleteCmd = &cobra.Command{
+	Use:   "delete <function_name>",
+	Short: "delete a function from Kubeless",
+	Long:  `delete a function from Kubeless`,
 	Run: func(cmd *cobra.Command, args []string) {
 		if len(args) != 1 {
-			logrus.Fatal("Need exactly one argument - route name")
+			logrus.Fatal("Need exactly one argument - function name")
 		}
-		ingName := args[0]
+		funcName := args[0]
 
 		ns, err := cmd.Flags().GetString("namespace")
 		if err != nil {
@@ -37,11 +40,18 @@ var routeDeleteCmd = &cobra.Command{
 			ns = utils.GetDefaultNamespace()
 		}
 
-		client := utils.GetClientOutOfCluster()
+		crdClient, err := utils.GetCRDClientOutOfCluster()
+		if err != nil {
+			logrus.Fatal(err)
+		}
 
-		err = utils.DeleteIngress(client, ingName, ns)
+		err = utils.DeleteK8sCustomResource(crdClient, funcName, ns)
 		if err != nil {
 			logrus.Fatal(err)
 		}
 	},
+}
+
+func init() {
+	deleteCmd.Flags().StringP("namespace", "", "", "Specify namespace for the function")
 }
