+++ cmd/kubeless/function/function_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package function
 
 import (
 	"archive/zip"
@@ -29,7 +29,6 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes/fake"
-	"k8s.io/client-go/pkg/apis/autoscaling/v2alpha1"
 )
 
 func TestParseLabel(t *testing.T) {
@@ -263,53 +262,5 @@ func TestGetFunctionDescription(t *testing.T) {
 	}
 	if result4.Spec.FunctionContentType != "base64+zip" {
 		t.Errorf("Should return base64+zip, received %s", result4.Spec.FunctionContentType)
-	}
-}
-
-func TestGetHorizontalAutoscaleDefinition(t *testing.T) {
-	var min, max int32
-	min = 1
-	max = 3
-	funcName := "test-autoscale"
-	ns := "default"
-	value := "10"
-	labels := map[string]string{
-		"foo": "bar",
-	}
-	metric := "cpu"
-	hpa, err := getHorizontalAutoscaleDefinition(funcName, ns, metric, min, max, value, labels)
-	if err != nil {
-		t.Fatalf("Unexpected error %v", err)
-	}
-	expectedMeta := metav1.ObjectMeta{
-		Name:      funcName,
-		Namespace: ns,
-		Labels:    labels,
-	}
-	if hpa.Spec.ScaleTargetRef.Name != funcName {
-		t.Fatalf("Creating wrong scale target name")
-	}
-	if !reflect.DeepEqual(expectedMeta, hpa.ObjectMeta) {
-		t.Errorf("Expected \n%v to be equal to \n%v", expectedMeta, hpa.ObjectMeta)
-	}
-	if *hpa.Spec.MinReplicas != min {
-		t.Errorf("Unexpected min replicas. Expecting %d got %d", min, *hpa.Spec.MinReplicas)
-	}
-	if hpa.Spec.MaxReplicas != max {
-		t.Errorf("Unexpected max replicas. Expecting %d got %d", max, hpa.Spec.MaxReplicas)
-	}
-	if hpa.Spec.Metrics[0].Type != v2alpha1.ResourceMetricSourceType ||
-		*hpa.Spec.Metrics[0].Resource.TargetAverageUtilization != int32(10) {
-		t.Error("Unexpected metric")
-	}
-
-	metric = "qps"
-	hpa, err = getHorizontalAutoscaleDefinition(funcName, ns, metric, min, max, value, labels)
-	if err != nil {
-		t.Fatalf("Unexpected error %v", err)
-	}
-	if hpa.Spec.Metrics[0].Type != v2alpha1.ObjectMetricSourceType ||
-		hpa.Spec.Metrics[0].Object.TargetValue.String() != "10" {
-		t.Error("Unexpected metric")
 	}
 }
