+++ cmd/kubeless/kubeless.go
@@ -20,6 +20,11 @@ package main
 import (
 	"os"
 
+	"github.com/kubeless/kubeless/cmd/kubeless/autoscale"
+	"github.com/kubeless/kubeless/cmd/kubeless/function"
+	"github.com/kubeless/kubeless/cmd/kubeless/route"
+	"github.com/kubeless/kubeless/cmd/kubeless/topic"
+	"github.com/kubeless/kubeless/cmd/kubeless/version"
 	"github.com/spf13/cobra"
 )
 
@@ -32,7 +37,7 @@ func newRootCmd() *cobra.Command {
 		Long:  globalUsage,
 	}
 
-	cmd.AddCommand(functionCmd, topicCmd, versionCmd, ingressCmd, autoscaleCmd)
+	cmd.AddCommand(function.FunctionCmd, topic.TopicCmd, version.VersionCmd, route.RouteCmd, autoscale.AutoscaleCmd)
 	return cmd
 }
