+++ cmd/kubeless/route/routeCreate.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package route
 
 import (
 	"github.com/kubeless/kubeless/pkg/spec"
