+++ cmd/kubeless/route/routeDelete.go
@@ -1,12 +1,9 @@
 /*
 Copyright (c) 2016-2017 Bitnami
-
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
-
     http://www.apache.org/licenses/LICENSE-2.0
-
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,26 +11,37 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package route
 
 import (
+	"github.com/kubeless/kubeless/pkg/utils"
+	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
-var topicCmd = &cobra.Command{
-	Use:   "topic SUBCOMMAND",
-	Short: "manage message topics in Kubeless",
-	Long:  `topic command allows user to list, create, delete topics on Kubeless`,
+var routeDeleteCmd = &cobra.Command{
+	Use:   "delete <name>",
+	Short: "delete a route from Kubeless",
+	Long:  `delete a route from Kubeless`,
 	Run: func(cmd *cobra.Command, args []string) {
-		cmd.Help()
+		if len(args) != 1 {
+			logrus.Fatal("Need exactly one argument - route name")
+		}
+		ingName := args[0]
+
+		ns, err := cmd.Flags().GetString("namespace")
+		if err != nil {
+			logrus.Fatal(err)
+		}
+		if ns == "" {
+			ns = utils.GetDefaultNamespace()
+		}
+
+		client := utils.GetClientOutOfCluster()
+
+		err = utils.DeleteIngress(client, ingName, ns)
+		if err != nil {
+			logrus.Fatal(err)
+		}
 	},
-}
-
-func init() {
-	cmds := []*cobra.Command{topicCreateCmd, topicDeleteCmd, topicListCmd, topicPublishCmd}
-
-	for _, cmd := range cmds {
-		topicCmd.AddCommand(cmd)
-		cmd.Flags().StringP("kafka-namespace", "", "kubeless", "Namespace where kafka-controller is deployed. It will default to 'kubeless'")
-	}
 }
